requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Base de datos
sqlalchemy==2.0.32
pymysql==1.1.1

# Configuración y validación
python-dotenv==1.0.1
pydantic==2.8.2
pydantic-settings==2.4.0

# LLM (Ollama Cloud vía HTTP)
requests==2.32.3
